:ambulance: Fix broken parsing of openai-timeout
Now that we have fixed the default value validation in glazed,
we uncover a bug that was previously hidden in parsing openai factory settings.

This is all a bit ugly, because we introduce a separate field to parse the timeout value in seconds and then convert it as a duration, which previously would be
used to set the parameter default value using SetDefaultsFromStruct.

Will do for now.